requirements.txt
# --- Web Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- Database ---
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9     # PostgreSQL driver

# --- Migrations ---
alembic==1.13.2

# --- Pydantic (Data Validation) ---
pydantic==2.9.2
pydantic-settings==2.6.1   # for managing settings via .env

# --- CORS (if you expose APIs to frontend) ---
starlette==0.38.5

# --- Optional / Utils ---
python-dotenv==1.0.1       # load environment variables from .env
